requirements.txt
# requirements.txt
Pillow==7.1.2
vk-api==11.8.0
bs4==4.9.1